net7.0: Run upgrade-assistant
Upgraded
 * FOCA
 * MetadataExtractCore
 * SearcherCore
 * Plugins API
 * Plugins Example

Also upgrades the project file format and the solutions (FOCA3_Pro and
Plugins API).

The upgrade to .NET 7.0 removed System.Web which needs to be addressed
in subsequent commits by migrating away from this deprecated API.

WinForms is also not part of .NET 7.0 proper but was moved to the
Windows Desktop Component, only available on Windows.

Some other reference assemblies are also missing from .NET 7.0 but
can be supplemented with the
Microsoft.NETFramework.ReferenceAssemblies package, even on non-Windows
platforms.